+++ app/src/main/kotlin/indi/dmzz_yyhyy/lightnovelreader/data/plugin/PluginUpdateInfo.kt
@@ -1,0 +1,7 @@
+package indi.dmzz_yyhyy.lightnovelreader.data.plugin
+
+data class PluginUpdateInfo(
+    val pluginId: String,
+    val versionName: String,
+    val storePlugin: StorePlugin
+)
